Add MecanumDrive.java for reference
Extended trajectoryFollowing until robot/target distance is within threshold (or timeout), instead of finish trajectory following as soon as trajectory duration is reached. If robot is able to obtain "pose" from another source (for example, AprilTag), set it as current robot position (assuming it is more accurate than the robot position from localizer). 

Note 1: if the other source can only turn the x/y/heading using robot coordination system, it need to be converted to the field coordination system before use.
Note 2: when using hub to run opencv to process image from uvc camera, delay is not ignorable. The adjustment information calculated is "out-dated" thus cannot be apply to robot current position. Robot position history has to be maintained, and adustment need to be offset-ed by the diff of current position and original position (when image is captured).